Restructre modules
