+++ src/index.ts
@@ -1,3 +1,4 @@
-export { default as Link } from "./Link";
+export * from "./Link";
 export * from "./Route";
-export * from "./Router";
+export * from "./BrowserRouter";
+export * from "./MemoryRouter";
